+++ resources/rules/uk.php
@@ -8,10 +8,16 @@ declare(strict_types=1);
  *
  * @return bool[]|array[]
  */
-return function (int $number, int $last_digit): array {
+return static function (int $number, int $last_digit): array {
     return [
-        'single' => $last_digit === 1 || $number === 0,
-        'plural' => $last_digit >= 2 && $last_digit < 5,
+        'single' => [
+            $number === 1,
+            $last_digit === 1 && $number >= 21,
+        ],
+        'plural' => [
+            $number >= 2 && $number < 5,
+            $number >= 22 && $last_digit >= 2 && $last_digit < 5,
+        ],
         'special' => [
             $number >= 5 && $number <= 20,
             $last_digit === 0,